reorganize server folder
