+++ server/utils/validation.js
@@ -12,15 +12,22 @@ export const validateProject = (instance) => {
   return ProjectDefinition.validate(instance);
 };
 
+//throws on error
 const idValidator = validators.id();
 
 /**
- * @description asserts valid ID, throws error if invalid or undefined
+ * @description validates an ID. callback called with error if has one, otherwise without any arguments
  * @param {uuid} id
+ * @param {function} callback
  */
-export const assertValidId = (id) => {
+export const assertValidId = (id, callback = () => {}) => {
   if (!id || typeof id !== 'string') {
-    throw new Error(errorNoIdProvided);
+    callback(errorNoIdProvided);
   }
-  idValidator(id);
+  try {
+    idValidator(id);
+    callback();
+  } catch (err) {
+    callback(errorIdInvalid);
+  }
 };
