+++ test/server/REST/clone.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import { Block as exampleBlock } from '../../schemas/_examples';
-import { set as dbSet } from '../../../server/utils/database';
+import { set as dbSet } from '../../../server/deprecated/database';
 import request from 'supertest';
 
 const devServer = require('../../../server/devServer');
