+++ test/server/REST/history.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import { Project as exampleProject } from '../../schemas/_examples';
-import { set as dbSet } from '../../../server/utils/database';
+import { set as dbSet } from '../../../server/deprecated/database';
 import request from 'supertest';
 import devServer from '../../../server/devServer';
