+++ test/server/REST/login.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import request from 'supertest';
-import { set as dbSet } from '../../../server/utils/database';
+import { set as dbSet } from '../../../server/deprecated/database';
 import { getSessionKey, login } from '../../../src/middleware/api';
 
 const devServer = require('../../../server/devServer');
